refactor(cache): rework resource cache manager internals
* Rework the implementation of ResourceCache
in order to use MemoryCache API. This new API
provides more options for expiration policies, which
will be configurable in a follow up modification.
This will be required, as the new approach gains readability
and avoids reflection by sacrificing memory as
it tracks named api resources with 2 different caches, one by name
and the other by Id. With a configurable expiration policy, this
disadvantage can be mitigated.
* Add dependencies required
for the in-memory caching implementation
and usage of runtime type inference